refactor(lynx): split AppBar hook and context
